+++ packages/ui/src/components/Button/Button.tsx
@@ -24,7 +24,7 @@ export const buttonVariants = cva(
         danger: 'bg-destructive text-destructive-foreground shadow-sm hover:bg-destructive/90',
         ghost: 'hover:bg-accent hover:text-accent-foreground',
         link: 'text-primary underline-offset-4 hover:underline',
-        outline: 'border border-input bg-background shadow-sm hover:bg-accent hover:text-accent-foreground',
+        outline: 'border border-input bg-inherit shadow-sm',
         primary: 'bg-primary text-primary-foreground shadow hover:bg-primary/90',
         secondary: 'bg-secondary border text-secondary-foreground shadow-sm hover:bg-secondary/80'
       }